requirements.txt
asgiref==3.7.2
Django==4.2.6
black==23.9.1
click==8.1.7
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.21
platformdirs==3.10.0
pytz==2023.3.post1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
